Initial check-in of entire project code
NOTE: If you are already using Collaborate in Unity, then checking code in and out of this Git will not work. You'll want to work in one content versioning system or the other, but not both because you'll create confusing conflicts. I'll try to write up a README with instructions.